Add .htaccess file, remove index.html
We want the user to drop straight into a PHP environment the moment that
they land on the page. In order to do so, you will have to enable the
mod_rewrite.so module in your httpd.conf file and restart the server.
The .htaccess file that I have included in this push just tells the
server to pull up index.php anytime a file that does not exist is
requested. The beauty of it is that you do not see the .php extension
when the page loads. Let me know if you need any help getting your
apache to work. We have to present on Tuesday, so i'm going to be moving
at an extremely very fast pace until then.